review 개선점(1)
1. js.key -> jsApiKey 로 변경
2. ReviewConstants,java 상수 클래스 제작 후 review 관련 파일에서 쓰던 상수들 다 정리했습니다.
3. CategoryProperties 에서 카테고리 테이블 생기면서 안쓰는 것들 다 없애고 categoryService 와 ServiceImpl 로 병합시켰습니다.
4. storeController 파일이 너무 기능이 많은 것 같아 파일을 좀 나눠봤습니다. 더 개선점 있으면 말해주십쇼
